requirements.txt
# AI & Machine Learning
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
langchain-openai==0.0.6
tiktoken==0.6.0
scikit-learn==1.4.1
statsmodels==0.14.1
joblib==1.3.2

# NLP
spacy==3.7.4
nltk==3.8.1

# Web Scraping & Automation
selenium==4.18.0
beautifulsoup4==4.12.3
scrapy==2.11.1
playwright==1.42.0
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.3

# Data Processing
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.2
motor==3.4.0

# API Frameworks
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Marketplace Integrations
python-amazon-sp-api==0.18.0
ebaysdk==2.2.0
facebook-sdk==3.1.0

# Payment Processing
stripe==8.2.0
paypal-checkout-serversdk==1.0.1

# Communication
twilio==9.0.1
sendgrid==6.11.0
python-telegram-bot==21.0.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
celery==5.3.6
apscheduler==3.10.4
tenacity==8.2.3
ratelimit==2.2.1
boto3==1.34.51
hvac==2.1.0

# Image Processing (for product authentication)
Pillow==10.2.0
opencv-python==4.9.0.80

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.20.0
sentry-sdk==1.40.5

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
faker==23.2.1

# Price Monitoring APIs
keepa==3.0.1
camelcamelcamel==0.1.0

# OCR for receipts/documents
pytesseract==0.3.10

# Browser automation
undetected-chromedriver==3.5.5

# Data validation
marshmallow==3.20.2
phonenumbers==8.13.31

# Scheduling
croniter==2.0.2

# Encryption
cryptography==42.0.3

# Price comparison
price-parser==0.3.4

# Proxy management
requests[socks]==2.31.0

